+++ src/pretalx/common/language.py
@@ -1,5 +1,8 @@
+from copy import copy
+
 from django.conf import global_settings, settings
 from django.utils.translation import get_language
+
 
 LANGUAGE_CODES_MAPPING = {
     language.lower(): language for language in settings.LANGUAGES_INFORMATION
@@ -13,7 +16,7 @@ LANGUAGE_NAMES.update(
 
 def get_language_information(lang: str):
     lang_key = LANGUAGE_CODES_MAPPING[lang.lower()]
-    information = settings.LANGUAGES_INFORMATION[lang_key]
+    information = copy(settings.LANGUAGES_INFORMATION[lang_key])
     information["code"] = lang
     return information